Create new pallet for LC verifier
