+++ parachain/pallets/verifier-lightclient/Cargo.toml
@@ -1,0 +1,91 @@
+[package]
+name = "pallet-verifier-lightclient"
+description = "Artemis Light Client Verifier Pallet"
+version = "0.1.1"
+edition = "2018"
+authors = ["Snowfork <contact@snowfork.com>"]
+repository = "https://github.com/Snowfork/polkadot-ethereum"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+hex = { package = "rustc-hex", version = "2.1.0", default-features = false }
+
+[dependencies.codec]
+default-features = false
+features = ["derive"]
+package = "parity-scale-codec"
+version = "1.3.4"
+
+[dev-dependencies.sp-keyring]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+tag = "v2.0.0-rc6"
+version = "2.0.0-rc6"
+
+[dependencies.frame-support]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+tag = "v2.0.0-rc6"
+version = "2.0.0-rc6"
+
+[dependencies.frame-system]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+tag = "v2.0.0-rc6"
+version = "2.0.0-rc6"
+
+[dependencies.sp-core]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+tag = "v2.0.0-rc6"
+version = "2.0.0-rc6"
+
+[dependencies.sp-io]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+tag = "v2.0.0-rc6"
+version = "2.0.0-rc6"
+
+[dependencies.sp-runtime]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+tag = "v2.0.0-rc6"
+version = "2.0.0-rc6"
+
+[dependencies.sp-std]
+default-features = false
+git = "https://github.com/paritytech/substrate.git"
+tag = "v2.0.0-rc6"
+version = "2.0.0-rc6"
+
+[dependencies.artemis-core]
+default-features = false
+path = "../../primitives/core"
+
+[dependencies.artemis-ethereum]
+default-features = false
+path = "../../primitives/ethereum"
+
+[dependencies.serde]
+package = "serde"
+optional = true
+version = "1.0.101"
+features = ["derive"]
+
+
+[features]
+default = ["std"]
+std = [
+    "serde/std",
+    "codec/std",
+    "frame-support/std",
+    "frame-system/std",
+    "sp-core/std",
+    "sp-io/std",
+    "sp-runtime/std",
+    "sp-std/std",
+    "artemis-core/std",
+    "artemis-ethereum/std",
+]
